@@ -2,9 +2,9 @@ source "https://rubygems.org"
 
 ruby File.read(File.join(File.dirname(__FILE__), ".ruby-version")).strip
 
-gem "rails", "~> 5.0.0"
+gem "rails", "~> 6.1.0"
 
-gem "attr_encrypted"
+gem "attr_encrypted", "~> 3.1.0"
 gem "bootstrap-sass", "~> 3.3.4"
 gem "bootstrap-scss"
 gem "carrierwave"
@@ -46,7 +46,7 @@ group :test do
   gem "database_cleaner"
   gem "faker"
   gem "launchy"
-  gem "show_me_the_cookies", '~> 5.0.1'
+  gem "show_me_the_cookies", '~> 6.0.0'
   gem "timecop"
 end
 
