@@ -1,7 +1,7 @@
 class DecryptedSecretsController < ApplicationController
   include RetrieveSecret
-  before_filter :retrieve_secret
-  before_filter :require_validated_email
+  before_action :retrieve_secret
+  before_action :require_validated_email
 
   def create
     begin
